Implement abandon manager. graphics pipeline now tightly bound to a single draw target. First pass write_combined upload logic. igpu now hosts all vulkan pipeline logic. Vulkan barrier manager now implements queue ownership transfers with semaphores. Vulkan job scopes now use explicit decorators to define weather they are writable. Vulkan queue now hosts ability to submit one time commands with semaphores.

+++ atmos/lib_vulkan/vulkan/shader/vulkan_attributes_decriptor.cpp
@@ -30,7 +30,7 @@ void vulkan_attributes_decriptor::reset( const attribute_indexer& indexer )
 		attribute_description->binding = source.compact_buffer_index;
 		attribute_description->location = (uint32_t)param.cfg().location;
 		attribute_description->format = param.cfg().vk.format;
-		attribute_description->offset = attr_cfg.offset;
+		attribute_description->offset = (uint32_t)attr_cfg.offset;
 	}
 
 	for ( uint32_t i = 0; i < indexer.buffer_count(); ++i )
